Allows to use a different clang version than the default system one
For some project using new c++ features it maybe useful to for the use
of a specific verison of clang.

This very small patch allow this use by setting up an environment variable.

The use of an environment variable is forced by tree.py that locates
the libclang statically and not using an argument. Otherwise an
argument to generate command would have been preffered.